svn path=/branches/header-work/; revision=47301

+++ dll/win32/hhctrl.ocx/hhctrl.c
@@ -302,7 +302,7 @@ int WINAPI doWinMain(HINSTANCE hInstance, LPSTR szCmdLine)
         }
         else
         {
-            FIXME("Unhandled HTML Help command line parameter! (%.*s)\n", space-szCmdLine, szCmdLine);
+            FIXME("Unhandled HTML Help command line parameter! (%.*s)\n", (int)(space-szCmdLine), szCmdLine);
             return 0;
         }
     }
